requirements.txt
websocket-client>=1.8.0
requests>=2.31.0
pandas>=2.1.0
numpy>=1.24.0
matplotlib>=3.8.0
freezegun==1.5.1
fastapi
uvicorn
python-socketio==5.11.0
python-engineio==4.9.0
websockets
